# api-specification.yaml
openapi: 3.0.0
info:
  title: Distributed Assessment Platform API
  version: 1.0.0
  description: |
    Scalable distributed assessment platform with automated notifications, reporting, and data export capabilities.
    
    ## Features
    - Role-based access control (Admin/Student)
    - Hierarchical content organization (Subject→Chapter→Quiz→Question)
    - Real-time quiz attempts with automatic scoring
    - Automated background jobs for daily reminders and monthly reports
    - Data export capabilities with CSV generation
    
    ## Authentication
    All endpoints require JWT Bearer token authentication except for login and registration.
    
  contact:
    name: Distributed Assessment Platform Support
    email: support@assessmentplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.assessmentplatform.com
    description: Production server
...
paths:
...
  /admin/quiz:
    get:
      tags: [Admin - Quiz]
      summary: Get all quizzes
      responses:
        '200':
          description: List of quizzes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizListResponse'
    post:
      tags: [Admin - Quiz]
      summary: Create a new quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
            example:
              chapter_id: 1
              name: Quiz 1
              description: First quiz
              start_date: "01-06-2025"
              end_date: "02-06-2025"
              duration: 30
      responses:
        '201':
          description: Quiz created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
  /admin/quiz/{quiz_id}:
    get:
      tags: [Admin - Quiz]
      summary: Get quiz by ID
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Quiz details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
    put:
      tags: [Admin - Quiz]
      summary: Update quiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
      responses:
        '200':
          description: Quiz updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
    delete:
      tags: [Admin - Quiz]
      summary: Delete quiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Quiz deleted successfully
  /admin/quiz/{quiz_id}/question:
    get:
      tags: [Admin - Quiz]
      summary: Get all questions for a specific quiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of questions for quiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
  /admin/quiz/search:
    get:
      tags: [Admin - Quiz]
      summary: Search quizzes
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        '200':
          description: Quiz search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizListResponse'

  # Admin - Question Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    QuizRequest:
      type: object
      required:
        - name
        - chapter_id
        - start_date
        - end_date
        - duration
      properties:
        name:
          type: string
          maxLength: 512
        chapter_id:
          type: integer
        description:
          type: string
          maxLength: 1024
        start_date:
          type: string
          format: date
          pattern: '^\d{2}-\d{2}-\d{4}$'
          example: "01-06-2025"
        end_date:
          type: string
          format: date
          pattern: '^\d{2}-\d{2}-\d{4}$'
          example: "02-06-2025"
        duration:
          type: integer
          minimum: 1
          description: Duration in minutes
    QuizResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        chapter_id:
          type: integer
        chapter_name:
          type: string
        subject_id:
          type: integer
        subject_name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        duration:
          type: integer
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        questions:
          type: integer
          description: Number of questions in the quiz
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuizListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuizResponse'

    # Question Schemas
...
    QuestionResponse:
      type: object
      properties:
        id:
          type: integer
        quiz_id:
          type: integer
        quiz_name:
          type: string
        chapter_id:
          type: integer
        chapter_name:
          type: string
        subject_id:
          type: integer
        subject_name:
          type: string
        question:
          type: string
        option1:
          type: string
        option2:
          type: string
        option3:
          type: string
        option4:
          type: string
        answer:
          type: integer
        marks:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuestionListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResponse'

    # Authentication Schemas
...
security:
  - BearerAuth: []